feature: add interfaces of port of C# library
At this commit, a port of the interfaces of the
C# library has been added. A lot of JavaDoc documentation
has been created.

"Database" in C# library has been renamed to "configuration"
in the Java part, as classes in there deal with configuration
and target sync systems not necesserily utilize a database
to store the configuration.
The same applies to the interfaces of the configuration system.
Their name was stripped of "Database".

Since there are no properties with implicit creation of
getters/setters in Java, as with C#, those "properties"
have been replace with getters only and their classes
have become interfaces, so that no setters are available
to the public.