feature: remove getRedirectUri() from ISettingsModel
Since this library does not handle OAuth directly but
authorization workflow must be implemented by the UI,
there is no use in any redirection URI.

In case a CLI version is created with some sort of library
support, then a special "localhost" URI must be used,
that is not configurable. The URI to use must be
well-known the the OAuth server. Otherwise it will be
rejected.

Therefore there is no use to pass the value to
this library by its consumer application.