refact: use sync API and remove use of interface "Future"
Big API refactoring to get rid of the asynchronous
API, that is based on java.util.concurrent.Future. In the end
it turned out that Java is too focused on synchronous execution
of tasks. Using Future would create a lot of micro threads
without real benefit, as the call to the HTTP RESTful API
of Smint.io does not (yet) support utilizing Futures.

The new focus is, to run the sync in a big thread as synchroizing
is not concurrently doable.