refactor: provide download File with ISmintIoBinary
The consumer of the binaries may request the binary data as a temporary File.
All temporary local files (if any) will be clean-up by default implementation
of `ISyncJob` (`DefaulSyncJob`).
So the consumer does not provide any path to temporary file
to the asset nor download himself.

No `InputStream` is used - directly using the stream from the HTTP response -
in order to avoid accidentally copying the stream data to memory. As some
binaries may contain multiple Gigabytes of data, it could affect the
memory management of the process and maybe crash the application.

This is a difference to the C# implementation, that need to be
ported to c# as well.