fix: (test) use Mockito instead of JMockit
Setting up JMockit to work with JUnit5 properly is cumbersome.
Working with Eclipse is nearly impossible, as JMockit removed
some auto detection feature of the library's required JVM
agent parameter. The developer thought it is to "hacky".

It seems to error prune to impose the necessity to all IDEs
to add the proper JVM agent settings to ALL JUnit run
configurations. Hence JMockit is exchanged for Mockito.

Since SpringFramework seem to prefer Mockito,
it replaces JMockit in this project, too.

see:

* https://stackoverflow.com/questions/2905735/jmockit-initialization-problem
* https://site.mockito.org
* https://stackoverflow.com/questions/40961057/how-to-use-mockito-with-junit5