refact: ISmintIo* use "Locale" in localized texts
Localized texts are Maps with languages as keys.
This commit changes the type of the keys to `Locale`, as
a locale is the standard Java way to specify languages.

The benefit to gain is the automatic handling of
all kinds of variants of the languages codes.

So there is no difference, whether the language is set
to be "en" or "eng", whereas `String` keys would differ
very much in that case. Upper case would make a difference too.