refact: make "isEditorialUse()" return "Boolean" object instead primitive type
A Boolean object may have a third value of `null`, indicating an unknown value.
Sometimes upstream providers do not set restrictions/allowence for non-editorial
use. So there is no decision, whether the content is for editorial use only
or not. Smint.io must not make any assumption in such cases. This is subject
to different laws in different countries.