feature: implement AuthTokenStorageWithNotify enforcing "notifyAll()" on IAuthTokenStorage
Using Java basic sync mechanism if Object's wait()/notifyAll() pair, is used
to notify every consumer of OAuth token data, that new data is available, in case
some instance depends on it and decides to wait for it.

Especially this is required right at the beginning of the sync process, when
no OAuth token data is available yet. Then there is a need to wait for the heavily
asynchronous process that involves external systems such as a user's browser and
the user's manual intervention. As soon as the user is involved, there is no fixed
time to wait for, so a wait/notify scheme must be used.