refact: make authorizer call notifyAll() on token storge, too
Make the default implementation of `ISmintIoOAuthAuthorizer`, which
is `SmintIoOAuthAuthorizer` call `notifyAll()` on the token storage.
This is necessary, as the token storage might not be called because
of invalid query parameters. So no token is to be stored and hence
the token storage would not call `notifyAll` on its own.