refact: introduce SystemBrowserAuthenticator start user interaction by local browser
The system browser authenticator opens up a browser with the URL as created
by the OAuth process in `ISmintIoOAuthAuthorizer.createAuthorizationUrl()`.
The user then will log in to Smint.io and then Smint.io autorization server
will redirect back to local server, with GET parameters holding the
OAuth code to request the OAuth access tokens.

Therefore a companion web server is being started, that will receive this
last request and OAuth code and then will pass the code the the
authenticator for analyzation. The authenticator then requests an OAuth
access token utilizing the code.