refact: remove wait/notify on IAuthTokenStorage - too complex
The scheme wait/notify was too complex to get it right. It was
introduced to help detection when the local HTTP server has
received OAuth URL parameters during the OAuth authorization
process. Then the web server is due to be shut down.

The new solution is much more elegant by not directly passing
the authorizer instance to the web server, but just a consumer
function. This consumer function will notify the caller of
the web server that OAuth process has been finished. But this
is up to the caller and much more flexible.

The new solution does not depend on external code, that is not
yet written or this library does not have control over.