refact: throw exception if invalid OAuth token data to be stored
In case the OAuth token data to be stored is invalid, storing is
rejected and an exception is thrown. This will help in detecting
the real cause of such errors.

Token data with "isSuccess()" being `true` but empty values is
rejected. Empty data with "isSuccess()" being `false` is accepted
though.