refact: implement a cache of 10 seconds into FileModelStorage
The cache is ignored in case the file has been altered. Then
the value is always being re-loaded. This situation is detected
by the last modified date of the file.

The cache time of 10 seconds is relatively short for humans, but
long for running programs. Its main purpose is to speed-up
repeated calls to "get()".