refact: remove downloading file from ISmintIoBinary
Downloading the file is performed by a download-provider
added to the next converted instance ISyncBinaryAsset.

Since downloading involves temporary file storage, it is
godd to have just one instance where to handle and clean-up.
There is no use in having two instances to be able to download
the file from Smint.io API.