fix: make series of chunk/batch loading of assets abort only at the end
Assets are loaded in batches/chunks from the Smint.io API server.
The synchronization jobs detects the end of the series of batches
as no assets are available anymore to be synchronized.

However, there are assets that are part of such a chunk that must
be ignored for synchronizing. So, if a batch contains of such
assets only, no assets to sync are created. The job handler
mistook this as the end of the list of chunks to sync. To avoid
this, a special check has been implemented to see, whether there
have been some assets available, which might then be ignored for
any sync.

Now the job continues the synchronization process.