refact: (build) remove "-SNAPSHOT" from manually generated jar names
Although using "SNAPSHOT" is a good habit for manually generated jar
files, it is not done with maven. Deployment of artifacts is performed
manually at the moment - using maven. So "-SNAPHOST" must not be used.

In order to achieve the same behaviour of maven and gradle, "-SNAPSHOT"
is removed from gradle build scripts, now.