refactor: introduce ISyncTargetDataFactory
A fresh instance of ISyncTarget is created for every run of
the sync job. The interface is very overloaded and has factory
functions as well. This is not necessary, as factory functions
are usually very static.

So the new interface ISyncTargetDataFactory is used to cover
the factory part, which is removed from ISyncTarget.