refact: (BREAKING) remove ID mapping from ISyncTarget
`ISyncTarget` is overloaded with functions and this refactoring
should reduce complexity a lot.

The mapping of Smint.io API IDs to sync target IDs is something
that is always necessary. So it is incorporated into the library
now. It helps to simplify implementation of `ISyncTarget` and
will provide a common mechanism for this mapping.

Everything that can be covered by this library will help to reduce
work load to implement `ISyncTarget`.