fix: #2 correct check for empty localized asset name
The previous check always failed, as it checked the converted
asset data and not the source asset data of the conversion.
Since the check is made during conversion, the converted asset
data does not hold any data at that very moment. So the check
will always fail and the asset name will not be converted but
get lost.