refact: #2 introduce wrapper classes for ISyncAsset, caching some data for sync job
The wrapper classes are caching some data for the implementation of ISyncJob.
Therefore the implementors of the SyncTarget instances do not need to
provide this data correctly. It will reduce potential source of error for
all data that is vital to ISyncJob implementation of DefaultSyncJob.

Despite using the wrapper, the ISyncTarget only received the unwrapped
instances.