fix: use English as fallback if requested import language has no meta-data values
For all meta data there exist localized texts and values. Nevertheless some
text and values might not be available in a specific language. With Smint.io
API a value for English will always be available!

If a sync target requests a language to synchronizes, for which there is no
value available, then the value of English is used - if available.

However, this is done only if English is not part of the requested sync
languages itself. In this case, the sync target must take care of using
English as default on its own.

There is a corner case, though. Keywords consist of multiple values for
a language. If at least a single keyword exists for a language, then there
is a value for the language available and thus no English values are used
instead. So there is no "merging" of English values into the values of
the requested language.