refact: (BREAKING) #2 use base class SyncAsset for assets
By using a single base class for both asset types of
Binary Asset and Compount Asset, it will be easier for
implementers. The base class already implements all functions
that handle transient data, so derived classes do not need to.

Additionally this eliminates a possible cause of error that
this transient data will be stored with the sync target -
as it effectively happend with the example application.