fix: pom.xml use fixed versions for submodule dependencies
It turned out that maven has problems with version ranges
on sub-module dependencies. On the other hand, all sub-modules
usually are changed at the same time. So most of the time
they will not work with older versions of other modules
they depend upon. So version number for all modules are
kept in sync now and use a fixed number.