refact: remove unused funcs to wait for queued jobs
The functions to wait for a scheduled job or be notified
about its execution, are making the job scheduler much
too complicated. These functions are not needed right now.

So they are removed.

Especially the notification scheme would impose the
necessity to use a new thread. But threads should not
(but are allowed) be used in the context of servlet
containers.