refact: loop the job execution queue until the queue is empty
Previously only a single item from the queue was executed.
This is insufficient as another item may be queued
while the first one is being executed. Thus waiting items
waited forever - until the next job was due to be executed.

Now the queue is executed until no more waiting items are
available in the queue.