fix: keep Pusher re-connecting on network error
Pusher's default settings make the Pusher library stop
reconnecting after maybe 3 minutes. So in case of a severe
network error of several minutes, loosing all connections,
Pusher's library will not reconnect.
The new settings make Pusher reconnect until stopped.

However, there remains a source of error, that is not handled
correctly. If the authentication token, that is used by
the Pusher library, expires in the meantime, then reconnecting
will faile without the chance of recovering.
This should not happen often and can be overcome by stopping the
synchronization and re-starting it. Then the Pusher connection
will be re-authorized and hence re-established.