fix: require implementation of ISyncJobDataStorage
Previously there was no enforcement to provide an
implementation of ISyncJobDataStorage. Since the lack
of such implementation means, its data will not be made
persistent, its provision is enforced now. This is vital
to make the synchronization always start with NEW assets
only and not sync all previous, old assets, too.
The job data contains a *Continuation UUID*, which is used
by the Smint.IO server to determine, which asset was the
last one that has been synced. So without this UUID,
synchronization always starts at the very first asset ever
purchased. So, it is necessary to store this UUID and pass
it to the server on every run.