fix:(auth) store empty auth token if refresh fails with "invalid_grant"
Error "invalid_grant" is used whenever the refresh token is not valid
anymore. By storing empty token data, re-authentication can be enforced.
Otherwise it is unclear whether the tokens are still valid or not.

This is not a breaking change but consumers of the library need to
make sure, they store "null" values as tokens in the persistent
storage too.