fix:(auth) fail fast if auth URL port is undefined
In case the OAuth callback URL does not contain a port specification,
because it is a `file://`-URL, failing is fast, now.
An exception is thrown to indicate the wrong, invalid usage of
the HTTP compagnion server, as this server only supports HTTP
and must listen on a network port.