feature(auth): new helper func to check for target of URL auth data
Multiple OAuth processes might be in progress. Authenticator helper
instances require to receive authentication data via URL parameters.
In case there are more concurrent operations in progress, it is an
effort to check for the proper helper instance of the authentication
data. The new function will help the receiving servlet to decide,
which authenticator is the target for the received URL data.