fix:(auth) deserialize OffsetDateTime for old nehaviour fix:(auth) deserialize OffsetDateTime for old serial data
Previously, no type adapter has been added to Gson for `OffsetDateTime`.
Thus a generic JavaScript object was created. Now that the proper
type adapter is in place, the old behaviour must still be maintained.
So a generic Gson with no type adaters is used to fetch the generic
object type adater, in order to deserialize the old data.

New serializations use proper available type adapter, which serialized
to a single string.