feature: (BREAK behaviour) skip Pusher channel if channel ID is invalid
Previously the Pusher service failed completely with an exception.
However, there is a legitimate and reasonable need to not force any
Pusher communication but solely rely on the scheduled synchronization
process. Therefore pusher service is skiped rather than failing on
invalid channel IDs.

This is a change in behaviour but no breaking change in API.