fix:(aut) read ID token from OAuth answer instead of missing it
The identity token is part of the OpenID token data. Since smint.io
platform makes use of OpenID there is such an identity token
available. Yet, since it is not part of the standard OAuth access
data class, it was not loaded previously.

Now this token is fetched from the access data and stored.
This should avoid an additionaly performed step to acquire
such an identity token.