refact:(JSON) use "SmintIoGsonProvider" wherever Gson is needed
Instead of creating a Gson instance and configure it at various
places, the Gson provider is used instead. It will make use
of a properly configured Gson intance.

We need to see, whether this will work with tests. It might be
questionabe, whether tests should use a plain Gson or the
application configured Gson. Since the tests should also cover
the classes working as part of the application, the application
Gson is preferred now. This might be reverted, once there is
a good reason why to use plain Gson for tests only.