fix:(CLAPI) ensure new access token is passed to API client
If accessing the Smint.io CLAPI failes with "authorization required"
a new access token is created. This token is stored in the token
storage. Previously it was not directly passed to the API client,
so there might be a slight chance that the token has not been
passed to the API client, although this should not be the case.
Since the supplier function to call the CLAPI was setting the token,
this should not have been an issue. Nevertheless the token will
be explicitly set to the API clients upon refreshing it.
Just to be sure.

This change chould not make a difference, but it is more explicit.
May using the new auto-refreh token storage, it will help, too.