fix:(CLAPI) catch HTTP "Unauthorized" APIException
Use `ThrowingSupplier.sneaky` instead of
`ThrowingSupplier.checked`!

The HTTP unauthorized APIException was not catched previously
because the supplier was created with `ThrowingSupplier.unchecked`.
It wrapped the `APIException` into a `WrappedException` class.
Hence the unauthorized access to CLAPI was undetected and the
sync process stopped with this exception.

Only the transaction API was affected, which is the most important
one. Calls to the other API wrappers were created properly with
`ThrowingSupplier.sneaky`.