fix:(auth) make auto-refresh auth storage fail silently on refresh
Auto refreshing auth token data is performed while providing the
data. This function does not indicate any throwing of exceptions.
So, any exception refreshing the auth data is ignored and just
logged. In such a case, the original and unchanged auth data
is returned, as if the refresh has never been tried.